+++ src/common/errors.rs
@@ -9,6 +9,9 @@
 //
 // This module defines errors that are used across the modules.
 
+#[cfg(feature = "python_lib")]
+use pyo3::{PyErr, exceptions::PyRuntimeError};
+
 use std::num::TryFromIntError;
 
 use crate::common::LabelElem;
@@ -112,6 +115,7 @@ pub enum TranslatorError {
     DepthNotFit(usize, usize),
     UndefinedWidth(u8),
     IOError(std::io::Error),
+    UndefinedFormat(String)
 }
 
 impl ExitErrorCode for TranslatorError {
@@ -120,6 +124,7 @@ impl ExitErrorCode for TranslatorError {
             TranslatorError::DepthNotFit(_, _) => 65,
             TranslatorError::UndefinedWidth(_) => 64,
             TranslatorError::IOError(_) => 73, // or 74 or 1
+            TranslatorError::UndefinedFormat(_) => 64, 
         }
     }
 }
@@ -136,7 +141,15 @@ impl std::fmt::Display for TranslatorError {
             TranslatorError::DepthNotFit(got, req) => write!(f, "Not enough addresses! {got} < {req}"),
             TranslatorError::UndefinedWidth(width) => write!(f, "Width {width} is not defined. Try using 8 or 32!"),
             TranslatorError::IOError(e) => write!(f, "{}", e),
+            TranslatorError::UndefinedFormat(format) => write!(f, "Format '{format}' is not defined. Try one of [\"mif\", \"raw\", \"debug\"]."),
         }
+    }
+}
+
+#[cfg(feature = "python_lib")]
+impl std::convert::From<TranslatorError> for PyErr {
+    fn from(err: TranslatorError) -> PyErr {
+        PyRuntimeError::new_err(err.to_string())
     }
 }
 
@@ -188,6 +201,12 @@ impl ExitErrorCode for LibraryError {
     }
 }
 
+#[cfg(feature = "python_lib")]
+impl std::convert::From<LibraryError> for PyErr {
+    fn from(err: LibraryError) -> PyErr {
+        PyRuntimeError::new_err(err.to_string())
+    }
+}
 pub enum ParserError {
     NoTextSection,
     CommonError(CommonError)